LPS-87529 Create DocumentDTOConverter and use it in DocumentResourceImpl
