+++ modules/apps/headless/headless-document-library/headless-document-library-impl/src/main/java/com/liferay/headless/document/library/internal/dto/v1_0/converter/DocumentDTOConverter.java
@@ -1,0 +1,194 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.headless.document.library.internal.dto.v1_0.converter;
+
+import com.liferay.adaptive.media.AMAttribute;
+import com.liferay.adaptive.media.AdaptiveMedia;
+import com.liferay.adaptive.media.image.finder.AMImageFinder;
+import com.liferay.adaptive.media.image.finder.AMImageQueryBuilder;
+import com.liferay.adaptive.media.image.mime.type.AMImageMimeTypeProvider;
+import com.liferay.adaptive.media.image.processor.AMImageAttribute;
+import com.liferay.adaptive.media.image.processor.AMImageProcessor;
+import com.liferay.asset.kernel.model.AssetTag;
+import com.liferay.asset.kernel.service.AssetCategoryLocalService;
+import com.liferay.asset.kernel.service.AssetTagLocalService;
+import com.liferay.document.library.kernel.model.DLFileEntry;
+import com.liferay.document.library.kernel.service.DLAppService;
+import com.liferay.document.library.util.DLURLHelper;
+import com.liferay.headless.document.library.dto.v1_0.AdaptedImage;
+import com.liferay.headless.document.library.dto.v1_0.Document;
+import com.liferay.headless.document.library.dto.v1_0.TaxonomyCategory;
+import com.liferay.headless.document.library.internal.dto.v1_0.util.AggregateRatingUtil;
+import com.liferay.headless.document.library.internal.dto.v1_0.util.CreatorUtil;
+import com.liferay.headless.web.experience.dto.v1_0.converter.DTOConverter;
+import com.liferay.headless.web.experience.dto.v1_0.converter.DTOConverterContext;
+import com.liferay.portal.kernel.comment.CommentManager;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.repository.model.FileVersion;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.ListUtil;
+import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.vulcan.util.TransformUtil;
+import com.liferay.ratings.kernel.service.RatingsStatsLocalService;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+import java.util.Optional;
+import java.util.stream.Stream;
+
+/**
+ * @author Rubén Pulido
+ */
+@Component(
+	property = "asset.entry.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
+	service = {DocumentDTOConverter.class, DTOConverter.class}
+)
+public class DocumentDTOConverter implements DTOConverter {
+
+	@Override
+	public Document toDTO(DTOConverterContext dtoConverterContext)
+		throws Exception {
+
+		FileEntry fileEntry = _dlAppService.getFileEntry(
+			dtoConverterContext.getResourcePrimKey());
+
+		FileVersion fileVersion = fileEntry.getFileVersion();
+
+		User user = _userService.getUserById(fileEntry.getUserId());
+
+		return new Document() {
+			{
+				adaptedImages = _getAdaptiveMedias(fileEntry);
+				aggregateRating = AggregateRatingUtil.toAggregateRating(
+					_ratingsStatsLocalService.fetchStats(
+						DLFileEntry.class.getName(),
+						fileEntry.getFileEntryId()));
+				contentUrl = _dlURLHelper.getPreviewURL(
+					fileEntry, fileVersion, null, "");
+				creator = CreatorUtil.toCreator(_portal, user);
+				dateCreated = fileEntry.getCreateDate();
+				dateModified = fileEntry.getModifiedDate();
+				description = fileEntry.getDescription();
+				encodingFormat = fileEntry.getMimeType();
+				fileExtension = fileEntry.getExtension();
+				folderId = fileEntry.getFolderId();
+				id = fileEntry.getFileEntryId();
+				keywords = ListUtil.toArray(
+					_assetTagLocalService.getTags(
+						DLFileEntry.class.getName(),
+						fileEntry.getFileEntryId()),
+					AssetTag.NAME_ACCESSOR);
+				numberOfComments = _commentManager.getCommentsCount(
+					DLFileEntry.class.getName(), fileEntry.getFileEntryId());
+				sizeInBytes = fileEntry.getSize();
+				taxonomyCategories = TransformUtil.transformToArray(
+					_assetCategoryLocalService.getCategories(
+						DLFileEntry.class.getName(),
+						fileEntry.getFileEntryId()),
+					assetCategory -> new TaxonomyCategory() {
+						{
+							taxonomyCategoryId = assetCategory.getCategoryId();
+							taxonomyCategoryName = assetCategory.getName();
+						}
+					},
+					TaxonomyCategory.class);
+				title = fileEntry.getTitle();
+			}
+		};
+	}
+
+	private AdaptedImage[] _getAdaptiveMedias(FileEntry fileEntry)
+		throws Exception {
+
+		if (!_amImageMimeTypeProvider.isMimeTypeSupported(
+				fileEntry.getMimeType())) {
+
+			return new AdaptedImage[0];
+		}
+
+		Stream<AdaptiveMedia<AMImageProcessor>> stream =
+			_amImageFinder.getAdaptiveMediaStream(
+				amImageQueryBuilder -> amImageQueryBuilder.forFileEntry(
+					fileEntry
+				).withConfigurationStatus(
+					AMImageQueryBuilder.ConfigurationStatus.ANY
+				).done());
+
+		return stream.map(
+			this::_toAdaptedImage
+		).toArray(
+			AdaptedImage[]::new
+		);
+	}
+
+	private <T, S> T _getValue(
+		AdaptiveMedia<S> adaptiveMedia, AMAttribute<S, T> amAttribute) {
+
+		Optional<T> valueOptional = adaptiveMedia.getValueOptional(amAttribute);
+
+		return valueOptional.orElse(null);
+	}
+
+	private AdaptedImage _toAdaptedImage(
+		AdaptiveMedia<AMImageProcessor> adaptiveMedia) {
+
+		return new AdaptedImage() {
+			{
+				contentUrl = String.valueOf(adaptiveMedia.getURI());
+				height = _getValue(
+					adaptiveMedia, AMImageAttribute.AM_IMAGE_ATTRIBUTE_HEIGHT);
+				resolutionName = _getValue(
+					adaptiveMedia,
+					AMAttribute.getConfigurationUuidAMAttribute());
+				sizeInBytes = _getValue(
+					adaptiveMedia, AMAttribute.getContentLengthAMAttribute());
+				width = _getValue(
+					adaptiveMedia, AMImageAttribute.AM_IMAGE_ATTRIBUTE_WIDTH);
+			}
+		};
+	}
+
+	@Reference
+	private AMImageFinder _amImageFinder;
+
+	@Reference
+	private AMImageMimeTypeProvider _amImageMimeTypeProvider;
+
+	@Reference
+	private AssetCategoryLocalService _assetCategoryLocalService;
+
+	@Reference
+	private AssetTagLocalService _assetTagLocalService;
+
+	@Reference
+	private CommentManager _commentManager;
+
+	@Reference
+	private DLAppService _dlAppService;
+
+	@Reference
+	private DLURLHelper _dlURLHelper;
+
+	@Reference
+	private Portal _portal;
+
+	@Reference
+	private RatingsStatsLocalService _ratingsStatsLocalService;
+
+	@Reference
+	private UserLocalService _userService;
+
+}
