+++ modules/apps/headless/headless-document-library/headless-document-library-impl/src/main/java/com/liferay/headless/document/library/internal/resource/v1_0/DocumentResourceImpl.java
@@ -14,34 +14,21 @@
 
 package com.liferay.headless.document.library.internal.resource.v1_0;
 
-import com.liferay.adaptive.media.AMAttribute;
-import com.liferay.adaptive.media.AdaptiveMedia;
-import com.liferay.adaptive.media.image.finder.AMImageFinder;
-import com.liferay.adaptive.media.image.finder.AMImageQueryBuilder;
-import com.liferay.adaptive.media.image.mime.type.AMImageMimeTypeProvider;
-import com.liferay.adaptive.media.image.processor.AMImageAttribute;
-import com.liferay.adaptive.media.image.processor.AMImageProcessor;
-import com.liferay.asset.kernel.model.AssetTag;
 import com.liferay.asset.kernel.service.AssetCategoryLocalService;
 import com.liferay.asset.kernel.service.AssetTagLocalService;
 import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.document.library.kernel.model.DLFolderConstants;
 import com.liferay.document.library.kernel.model.DLVersionNumberIncrease;
 import com.liferay.document.library.kernel.service.DLAppService;
-import com.liferay.document.library.util.DLURLHelper;
 import com.liferay.headless.common.spi.service.context.ServiceContextUtil;
-import com.liferay.headless.document.library.dto.v1_0.AdaptedImage;
 import com.liferay.headless.document.library.dto.v1_0.Document;
-import com.liferay.headless.document.library.dto.v1_0.TaxonomyCategory;
-import com.liferay.headless.document.library.internal.dto.v1_0.util.AggregateRatingUtil;
-import com.liferay.headless.document.library.internal.dto.v1_0.util.CreatorUtil;
+import com.liferay.headless.document.library.internal.dto.v1_0.converter.DocumentDTOConverter;
 import com.liferay.headless.document.library.internal.odata.entity.v1_0.DocumentEntityModel;
 import com.liferay.headless.document.library.resource.v1_0.DocumentResource;
+import com.liferay.headless.web.experience.dto.v1_0.converter.DefaultDTOConverterContext;
 import com.liferay.petra.function.UnsafeConsumer;
-import com.liferay.portal.kernel.comment.CommentManager;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.repository.model.FileVersion;
 import com.liferay.portal.kernel.repository.model.Folder;
 import com.liferay.portal.kernel.search.BooleanClauseOccur;
 import com.liferay.portal.kernel.search.BooleanQuery;
@@ -50,11 +37,8 @@ import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.search.filter.BooleanFilter;
 import com.liferay.portal.kernel.search.filter.Filter;
 import com.liferay.portal.kernel.search.filter.TermFilter;
-import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.odata.entity.EntityModel;
 import com.liferay.portal.vulcan.multipart.BinaryFile;
 import com.liferay.portal.vulcan.multipart.MultipartBody;
@@ -62,11 +46,8 @@ import com.liferay.portal.vulcan.pagination.Page;
 import com.liferay.portal.vulcan.pagination.Pagination;
 import com.liferay.portal.vulcan.resource.EntityModelResource;
 import com.liferay.portal.vulcan.util.SearchUtil;
-import com.liferay.ratings.kernel.service.RatingsEntryLocalService;
-import com.liferay.ratings.kernel.service.RatingsStatsLocalService;
 
 import java.util.Optional;
-import java.util.stream.Stream;
 
 import javax.ws.rs.BadRequestException;
 import javax.ws.rs.core.Context;
@@ -213,9 +194,7 @@ public class DocumentResourceImpl
 					null
 				)));
 
-		return _toDocument(
-			fileEntry, fileEntry.getFileVersion(),
-			_userService.getUserById(fileEntry.getUserId()));
+		return _toDocument(fileEntry);
 	}
 
 	@Override
@@ -296,9 +275,7 @@ public class DocumentResourceImpl
 					Document.ViewableBy.OWNER.getValue()
 				)));
 
-		return _toDocument(
-			fileEntry, fileEntry.getFileVersion(),
-			_userService.getUserById(fileEntry.getUserId()));
+		return _toDocument(fileEntry);
 	}
 
 	private Document _addDocument(
@@ -348,33 +325,7 @@ public class DocumentResourceImpl
 					Document.ViewableBy.OWNER.getValue()
 				)));
 
-		return _toDocument(
-			fileEntry, fileEntry.getFileVersion(),
-			_userService.getUserById(fileEntry.getUserId()));
-	}
-
-	private AdaptedImage[] _getAdaptiveMedias(FileEntry fileEntry)
-		throws Exception {
-
-		if (!_amImageMimeTypeProvider.isMimeTypeSupported(
-				fileEntry.getMimeType())) {
-
-			return new AdaptedImage[0];
-		}
-
-		Stream<AdaptiveMedia<AMImageProcessor>> stream =
-			_amImageFinder.getAdaptiveMediaStream(
-				amImageQueryBuilder -> amImageQueryBuilder.forFileEntry(
-					fileEntry
-				).withConfigurationStatus(
-					AMImageQueryBuilder.ConfigurationStatus.ANY
-				).done());
-
-		return stream.map(
-			this::_toAdaptedImage
-		).toArray(
-			AdaptedImage[]::new
-		);
+		return _toDocument(fileEntry);
 	}
 
 	private Page<Document> _getDocumentsPage(
@@ -395,91 +346,12 @@ public class DocumentResourceImpl
 			sorts);
 	}
 
-	private <T, S> T _getValue(
-		AdaptiveMedia<S> adaptiveMedia, AMAttribute<S, T> amAttribute) {
-
-		Optional<T> valueOptional = adaptiveMedia.getValueOptional(amAttribute);
-
-		return valueOptional.orElse(null);
-	}
-
-	private AdaptedImage _toAdaptedImage(
-		AdaptiveMedia<AMImageProcessor> adaptiveMedia) {
-
-		return new AdaptedImage() {
-			{
-				contentUrl = String.valueOf(adaptiveMedia.getURI());
-				height = _getValue(
-					adaptiveMedia, AMImageAttribute.AM_IMAGE_ATTRIBUTE_HEIGHT);
-				resolutionName = _getValue(
-					adaptiveMedia,
-					AMAttribute.getConfigurationUuidAMAttribute());
-				sizeInBytes = _getValue(
-					adaptiveMedia, AMAttribute.getContentLengthAMAttribute());
-				width = _getValue(
-					adaptiveMedia, AMImageAttribute.AM_IMAGE_ATTRIBUTE_WIDTH);
-			}
-		};
-	}
-
 	private Document _toDocument(FileEntry fileEntry) throws Exception {
-		return _toDocument(
-			fileEntry, fileEntry.getFileVersion(),
-			_userLocalService.getUserById(fileEntry.getUserId()));
-	}
-
-	private Document _toDocument(
-			FileEntry fileEntry, FileVersion fileVersion, User user)
-		throws Exception {
-
-		return new Document() {
-			{
-				adaptedImages = _getAdaptiveMedias(fileEntry);
-				aggregateRating = AggregateRatingUtil.toAggregateRating(
-					_ratingsStatsLocalService.fetchStats(
-						DLFileEntry.class.getName(),
-						fileEntry.getFileEntryId()));
-				contentUrl = _dlURLHelper.getPreviewURL(
-					fileEntry, fileVersion, null, "");
-				creator = CreatorUtil.toCreator(_portal, user);
-				dateCreated = fileEntry.getCreateDate();
-				dateModified = fileEntry.getModifiedDate();
-				description = fileEntry.getDescription();
-				encodingFormat = fileEntry.getMimeType();
-				fileExtension = fileEntry.getExtension();
-				folderId = fileEntry.getFolderId();
-				id = fileEntry.getFileEntryId();
-				keywords = ListUtil.toArray(
-					_assetTagLocalService.getTags(
-						DLFileEntry.class.getName(),
-						fileEntry.getFileEntryId()),
-					AssetTag.NAME_ACCESSOR);
-				numberOfComments = _commentManager.getCommentsCount(
-					DLFileEntry.class.getName(), fileEntry.getFileEntryId());
-				sizeInBytes = fileEntry.getSize();
-				taxonomyCategories = transformToArray(
-					_assetCategoryLocalService.getCategories(
-						DLFileEntry.class.getName(),
-						fileEntry.getFileEntryId()),
-					assetCategory -> new TaxonomyCategory() {
-						{
-							taxonomyCategoryId = assetCategory.getCategoryId();
-							taxonomyCategoryName = assetCategory.getName();
-						}
-					},
-					TaxonomyCategory.class);
-				title = fileEntry.getTitle();
-			}
-		};
+		return _documentDTOConverter.toDTO(
+			new DefaultDTOConverterContext(null, fileEntry.getFileEntryId()));
 	}
 
 	private static final EntityModel _entityModel = new DocumentEntityModel();
-
-	@Reference
-	private AMImageFinder _amImageFinder;
-
-	@Reference
-	private AMImageMimeTypeProvider _amImageMimeTypeProvider;
 
 	@Reference
 	private AssetCategoryLocalService _assetCategoryLocalService;
@@ -488,30 +360,12 @@ public class DocumentResourceImpl
 	private AssetTagLocalService _assetTagLocalService;
 
 	@Reference
-	private CommentManager _commentManager;
-
-	@Reference
 	private DLAppService _dlAppService;
 
 	@Reference
-	private DLURLHelper _dlURLHelper;
-
-	@Reference
-	private Portal _portal;
-
-	@Reference
-	private RatingsEntryLocalService _ratingsEntryLocalService;
-
-	@Reference
-	private RatingsStatsLocalService _ratingsStatsLocalService;
+	private DocumentDTOConverter _documentDTOConverter;
 
 	@Context
 	private User _user;
-
-	@Reference
-	private UserLocalService _userLocalService;
-
-	@Reference
-	private UserLocalService _userService;
 
 }
